pyproject.toml
[project]
name = "git-ml-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Kyle Newman", email = "kyle@thenewmans.com.au" }]
dependencies = [
  "PyDriller>=2.7",
  "GitPython>=3.1.44",
  "fastapi[standard]>=0.115.11",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "None" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[dependency-groups]
dev = ["pytest>=8.3.5"]
